LPS-87529 Create BlogPostingDTOConverter and use it in BlogPostingResourceImpl
