+++ modules/apps/headless/headless-collaboration/headless-collaboration-impl/src/main/java/com/liferay/headless/collaboration/internal/dto/v1_0/converter/BlogPostingDTOConverter.java
@@ -1,0 +1,149 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.headless.collaboration.internal.dto.v1_0.converter;
+
+import com.liferay.asset.kernel.model.AssetTag;
+import com.liferay.asset.kernel.service.AssetCategoryLocalService;
+import com.liferay.asset.kernel.service.AssetTagLocalService;
+import com.liferay.blogs.model.BlogsEntry;
+import com.liferay.blogs.service.BlogsEntryService;
+import com.liferay.document.library.kernel.service.DLAppService;
+import com.liferay.document.library.util.DLURLHelper;
+import com.liferay.headless.collaboration.dto.v1_0.BlogPosting;
+import com.liferay.headless.collaboration.dto.v1_0.Image;
+import com.liferay.headless.collaboration.dto.v1_0.TaxonomyCategory;
+import com.liferay.headless.collaboration.internal.dto.v1_0.util.AggregateRatingUtil;
+import com.liferay.headless.collaboration.internal.dto.v1_0.util.CreatorUtil;
+import com.liferay.headless.web.experience.dto.v1_0.converter.DTOConverter;
+import com.liferay.headless.web.experience.dto.v1_0.converter.DTOConverterContext;
+import com.liferay.portal.kernel.comment.CommentManager;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.repository.model.FileEntry;
+import com.liferay.portal.kernel.service.UserLocalService;
+import com.liferay.portal.kernel.util.ListUtil;
+import com.liferay.portal.kernel.util.Portal;
+import com.liferay.portal.vulcan.util.TransformUtil;
+import com.liferay.ratings.kernel.service.RatingsStatsLocalService;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+import javax.ws.rs.core.Context;
+
+/**
+ * @author Rubén Pulido
+ */
+@Component(
+	property = "asset.entry.class.name=com.liferay.blogs.model.BlogsEntry",
+	service = {BlogPostingDTOConverter.class, DTOConverter.class}
+)
+public class BlogPostingDTOConverter implements DTOConverter {
+
+	@Override
+	public BlogPosting toDTO(DTOConverterContext dtoConverterContext)
+		throws Exception {
+
+		BlogsEntry blogsEntry = _blogsEntryService.getEntry(
+			dtoConverterContext.getResourcePrimKey());
+
+		return new BlogPosting() {
+			{
+				alternativeHeadline = blogsEntry.getSubtitle();
+				aggregateRating = AggregateRatingUtil.toAggregateRating(
+					_ratingsStatsLocalService.fetchStats(
+						BlogsEntry.class.getName(), blogsEntry.getEntryId()));
+				articleBody = blogsEntry.getContent();
+				contentSpaceId = blogsEntry.getGroupId();
+				creator = CreatorUtil.toCreator(
+					_portal, _userLocalService.getUser(blogsEntry.getUserId()));
+				dateCreated = blogsEntry.getCreateDate();
+				dateModified = blogsEntry.getModifiedDate();
+				datePublished = blogsEntry.getDisplayDate();
+				description = blogsEntry.getDescription();
+				encodingFormat = "text/html";
+				friendlyUrlPath = blogsEntry.getUrlTitle();
+				headline = blogsEntry.getTitle();
+				id = blogsEntry.getEntryId();
+				image = _getImage(blogsEntry);
+				keywords = ListUtil.toArray(
+					_assetTagLocalService.getTags(
+						BlogsEntry.class.getName(), blogsEntry.getEntryId()),
+					AssetTag.NAME_ACCESSOR);
+				numberOfComments = _commentManager.getCommentsCount(
+					BlogsEntry.class.getName(), blogsEntry.getEntryId());
+				taxonomyCategories = TransformUtil.transformToArray(
+					_assetCategoryLocalService.getCategories(
+						BlogsEntry.class.getName(), blogsEntry.getEntryId()),
+					assetCategory -> new TaxonomyCategory() {
+						{
+							taxonomyCategoryId = assetCategory.getCategoryId();
+							taxonomyCategoryName = assetCategory.getName();
+						}
+					},
+					TaxonomyCategory.class);
+			}
+		};
+	}
+
+	private Image _getImage(BlogsEntry blogsEntry) throws Exception {
+		long coverImageFileEntryId = blogsEntry.getCoverImageFileEntryId();
+
+		if (coverImageFileEntryId == 0) {
+			return null;
+		}
+
+		FileEntry fileEntry = _dlAppService.getFileEntry(coverImageFileEntryId);
+
+		return new Image() {
+			{
+				caption = blogsEntry.getCoverImageCaption();
+				contentUrl = _dlURLHelper.getPreviewURL(
+					fileEntry, fileEntry.getFileVersion(), null, "", false,
+					false);
+				imageId = coverImageFileEntryId;
+			}
+		};
+	}
+
+	@Reference
+	private AssetCategoryLocalService _assetCategoryLocalService;
+
+	@Reference
+	private AssetTagLocalService _assetTagLocalService;
+
+	@Reference
+	private BlogsEntryService _blogsEntryService;
+
+	@Reference
+	private CommentManager _commentManager;
+
+	@Reference
+	private DLAppService _dlAppService;
+
+	@Reference
+	private DLURLHelper _dlURLHelper;
+
+	@Reference
+	private Portal _portal;
+
+	@Reference
+	private RatingsStatsLocalService _ratingsStatsLocalService;
+
+	@Context
+	private User _user;
+
+	@Reference
+	private UserLocalService _userLocalService;
+
+}
