+++ modules/apps/headless/headless-collaboration/headless-collaboration-impl/src/main/java/com/liferay/headless/collaboration/internal/resource/v1_0/BlogPostingResourceImpl.java
@@ -14,33 +14,24 @@
 
 package com.liferay.headless.collaboration.internal.resource.v1_0;
 
-import com.liferay.asset.kernel.model.AssetTag;
-import com.liferay.asset.kernel.service.AssetCategoryLocalService;
-import com.liferay.asset.kernel.service.AssetTagLocalService;
 import com.liferay.blogs.model.BlogsEntry;
 import com.liferay.blogs.service.BlogsEntryService;
 import com.liferay.document.library.kernel.service.DLAppService;
-import com.liferay.document.library.util.DLURLHelper;
 import com.liferay.headless.collaboration.dto.v1_0.BlogPosting;
 import com.liferay.headless.collaboration.dto.v1_0.Image;
-import com.liferay.headless.collaboration.dto.v1_0.TaxonomyCategory;
-import com.liferay.headless.collaboration.internal.dto.v1_0.util.AggregateRatingUtil;
-import com.liferay.headless.collaboration.internal.dto.v1_0.util.CreatorUtil;
+import com.liferay.headless.collaboration.internal.dto.v1_0.converter.BlogPostingDTOConverter;
 import com.liferay.headless.collaboration.internal.odata.entity.v1_0.BlogPostingEntityModel;
 import com.liferay.headless.collaboration.resource.v1_0.BlogPostingResource;
 import com.liferay.headless.common.spi.service.context.ServiceContextUtil;
-import com.liferay.portal.kernel.comment.CommentManager;
+import com.liferay.headless.web.experience.dto.v1_0.converter.DefaultDTOConverterContext;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.search.Field;
 import com.liferay.portal.kernel.search.Sort;
 import com.liferay.portal.kernel.search.filter.Filter;
-import com.liferay.portal.kernel.service.UserLocalService;
 import com.liferay.portal.kernel.servlet.taglib.ui.ImageSelector;
 import com.liferay.portal.kernel.util.FileUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.ListUtil;
-import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.odata.entity.EntityModel;
 import com.liferay.portal.vulcan.pagination.Page;
@@ -54,14 +45,16 @@ import java.time.LocalDateTime;
 
 import java.util.Date;
 import java.util.Optional;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+import org.osgi.service.component.annotations.ServiceScope;
 
 import javax.ws.rs.BadRequestException;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MultivaluedMap;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-import org.osgi.service.component.annotations.ServiceScope;
+import java.time.LocalDateTime;
+import java.util.Date;
+import java.util.Optional;
 
 /**
  * @author Javier Gamarra
@@ -197,26 +190,6 @@ public class BlogPostingResourceImpl
 		blogPosting.setNumberOfComments((Number)null);
 	}
 
-	private Image _getImage(BlogsEntry blogsEntry) throws Exception {
-		long coverImageFileEntryId = blogsEntry.getCoverImageFileEntryId();
-
-		if (coverImageFileEntryId == 0) {
-			return null;
-		}
-
-		FileEntry fileEntry = _dlAppService.getFileEntry(coverImageFileEntryId);
-
-		return new Image() {
-			{
-				caption = blogsEntry.getCoverImageCaption();
-				contentUrl = _dlURLHelper.getPreviewURL(
-					fileEntry, fileEntry.getFileVersion(), null, "", false,
-					false);
-				imageId = coverImageFileEntryId;
-			}
-		};
-	}
-
 	private ImageSelector _getImageSelector(Long imageId) {
 		if ((imageId == null) || (imageId == 0)) {
 			return new ImageSelector();
@@ -237,76 +210,22 @@ public class BlogPostingResourceImpl
 	}
 
 	private BlogPosting _toBlogPosting(BlogsEntry blogsEntry) throws Exception {
-		return new BlogPosting() {
-			{
-				alternativeHeadline = blogsEntry.getSubtitle();
-				aggregateRating = AggregateRatingUtil.toAggregateRating(
-					_ratingsStatsLocalService.fetchStats(
-						BlogsEntry.class.getName(), blogsEntry.getEntryId()));
-				articleBody = blogsEntry.getContent();
-				contentSpaceId = blogsEntry.getGroupId();
-				creator = CreatorUtil.toCreator(
-					_portal, _userLocalService.getUser(blogsEntry.getUserId()));
-				dateCreated = blogsEntry.getCreateDate();
-				dateModified = blogsEntry.getModifiedDate();
-				datePublished = blogsEntry.getDisplayDate();
-				description = blogsEntry.getDescription();
-				encodingFormat = "text/html";
-				friendlyUrlPath = blogsEntry.getUrlTitle();
-				headline = blogsEntry.getTitle();
-				id = blogsEntry.getEntryId();
-				image = _getImage(blogsEntry);
-				keywords = ListUtil.toArray(
-					_assetTagLocalService.getTags(
-						BlogsEntry.class.getName(), blogsEntry.getEntryId()),
-					AssetTag.NAME_ACCESSOR);
-				numberOfComments = _commentManager.getCommentsCount(
-					BlogsEntry.class.getName(), blogsEntry.getEntryId());
-				taxonomyCategories = transformToArray(
-					_assetCategoryLocalService.getCategories(
-						BlogsEntry.class.getName(), blogsEntry.getEntryId()),
-					assetCategory -> new TaxonomyCategory() {
-						{
-							taxonomyCategoryId = assetCategory.getCategoryId();
-							taxonomyCategoryName = assetCategory.getName();
-						}
-					},
-					TaxonomyCategory.class);
-			}
-		};
+		return _blogPostingDTOConverter.toDTO(
+			new DefaultDTOConverterContext(
+				contextAcceptLanguage.getPreferredLocale(),
+				blogsEntry.getEntryId()));
 	}
 
 	private static final EntityModel _entityModel =
 		new BlogPostingEntityModel();
 
 	@Reference
-	private AssetCategoryLocalService _assetCategoryLocalService;
-
-	@Reference
-	private AssetTagLocalService _assetTagLocalService;
+	private BlogPostingDTOConverter _blogPostingDTOConverter;
 
 	@Reference
 	private BlogsEntryService _blogsEntryService;
 
 	@Reference
-	private CommentManager _commentManager;
-
-	@Reference
 	private DLAppService _dlAppService;
-
-	@Reference
-	private DLURLHelper _dlURLHelper;
-
-	@Reference
-	private Portal _portal;
-
-	@Reference
-	private RatingsStatsLocalService _ratingsStatsLocalService;
-
-	@Context
-	private User _user;
-
-	@Reference
-	private UserLocalService _userLocalService;
 
 }
